fix frozen time on website

+++ src/ranking/ranking_helper.hh
@@ -30,7 +30,7 @@ namespace shiro::ranking::helper {
 
     int32_t get_leaderboard_position(uint8_t mode, std::string username, bool is_relax);
 
-    std::string get_leaderboard_user(uint8_t mode, int32_t pos, bool is_relax);
+    std::string get_leaderboard_user(uint8_t mode, size_t pos, bool is_relax);
 
     int16_t get_pp_for_user(uint8_t mode, std::string username, bool is_relax);
